backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.28.0
SQLAlchemy==2.0.29
asyncpg==0.29.0
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic-settings==2.2.1
jinja2==3.1.3
weasyprint==61.2
resend==0.8.0
python-dotenv==1.0.1
pytest==8.1.1
httpx==0.27.0
